Allow to hook history type info
Change `ExtensionHook` to keep a list of types.
Change `ExtensionLoader` to add/remove the types when the extension is
added/removed.
Change `HrefTypeInfo` to allow to dynamically add/remove the types.
Move the manual type to `ExtensionManualRequestEditor` to not have
other parts of the codebase depend on its classes, which will be
superseded.
Remove usage of the type in `AbstractHistoryReferencesTableEntry` use
hardcoded value instead.

Signed-off-by: thc202 <thc202@gmail.com>